Reorganise email feature
Move the feature into its own folder and then split into separate scenarios. Hopefully this makes it easier to work with and stops the steps file from getting out of control.
